remove duplicate formatters
