+++ packages/web/app/components/eto/overview/EtoOverviewThumbnail/EtoOverviewThumbnail.tsx
@@ -1,5 +1,5 @@
 import { Eur } from "@neufund/design-system";
-import { EAbbreviatedNumberOutputFormat, XOR, ERoundingMode } from "@neufund/shared";
+import { EAbbreviatedNumberOutputFormat, ERoundingMode, XOR } from "@neufund/shared";
 import * as React from "react";
 import { FormattedMessage } from "react-intl-phraseapp";
 import { branch, compose, renderComponent } from "recompose";
